Modified quickstart to discourage forking workflow
To keep config repos lightweight and easy to navigate, it's much nicer to
employ submodules to include the SALVE source. Rather than encouraging forking
the repo and layer (as was the previous recommendation, on account of the
SALVE_ROOT dir and so forth), we now encourage the use of submodules in spite
of the difficulties that they introduce. They better describe a dependent
repository than a fork.